requirements.txt
absl-py==2.1.0
accelerate==1.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
annotated-types==0.7.0
arguments==76
astunparse==1.6.3
attrs==24.2.0
beautifulsoup4==4.12.3
blis==1.1.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
cloudpathlib==0.20.0
confection==0.1.5
consoleprinter==95
cymem==2.0.10
datasets==3.2.0
dill==0.3.8
einops==0.8.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
filelock==3.16.1
flatbuffers==24.12.23
frozenlist==1.5.0
fsspec==2024.9.0
future==1.0.0
gast==0.6.0
gdown==5.2.0
gensim==4.3.2
google-pasta==0.2.0
grpcio==1.68.0
h5py==3.12.1
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
keras==3.7.0
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
lightning-utilities==0.11.9
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.11
namex==0.0.8
networkx==3.4.2
nltk==3.9.1
numpy==2.0.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.4.127
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
peft==0.8.2
pillow==11.0.0
preshed==3.0.9
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.1
pyarrow==18.1.0
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
pyphen==0.17.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-lightning==1.8.5.post0
pytorch-metric-learning==2.7.0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==2.4.0
sentencepiece==0.2.0
setuptools==75.6.0
shellingham==1.5.4
simplification==0.7.13
six==1.16.0
smart-open==7.1.0
soupsieve==2.6
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-syllables==3.0.2
srsly==2.5.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.18.0
termcolor==2.5.0
terminaltables==3.1.10
tf_keras==2.18.0
thinc==8.3.3
threadpoolctl==3.5.0
tokenizers==0.20.3
tomli==2.2.1
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchmetrics==1.6.0
torchvision==0.20.1+cu118
tqdm==4.67.1
transformers==4.46.3
triton==3.1.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
xxhash==3.5.0
yarl==1.18.0
